Refactored pseudo modifiers for Path of Exile 1 and 2.

+++ src/Sidekick.Apis.Poe/Parser/Pseudo/PseudoPattern.cs
@@ -5,10 +5,10 @@ namespace Sidekick.Apis.Poe.Parser.Pseudo;
 public class PseudoPattern
 (
     Regex regex,
-    int multiplier = 1
+    double multiplier = 1
 )
 {
     public Regex Pattern { get; set; } = regex;
 
-    public int Multiplier { get; set; } = multiplier;
+    public double Multiplier { get; set; } = multiplier;
 }
